fixed errors for new version of boost 1.67 and posix_time requiring casting to long

+++ host/examples/tx_samples_from_file.cpp
@@ -197,7 +197,7 @@ int UHD_SAFE_MAIN(int argc, char *argv[]){
         else throw std::runtime_error("Unknown type " + type);
 
         if(repeat and delay > 0.0) {
-            boost::this_thread::sleep(boost::posix_time::milliseconds(delay));
+            boost::this_thread::sleep(boost::posix_time::milliseconds((long)delay));
             std::this_thread::sleep_for(
                 std::chrono::milliseconds(int64_t(delay*1000))
             );
